Fix blurb-details swap
There are a handful of events (mostly birthdays) which only have text in EITHER "Blurb" or "Details". In all but two cases the text was under "Details". I'm moving the text in these two events from "Blurb" to "Details" so everything is consistent. We found that when non-English translations don't match the English it causes a bug, so consistency on this seems best.